Removed internal namespace from RangeSelector.
